requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pytest==8.0.0
pytest-cov==4.1.0
flake8==7.0.0
isort==5.13.2 